+++ angular-view/src/app/app.global-variable.ts
@@ -1,0 +1,7 @@
+
+export const URL = Object.freeze({
+    BASE:  '/',
+    AUTH_PROFILE:  'auth/profile',
+    AUTH_GOOGLE:  'auth/google',
+    API_WORDITEMS:  'api/word-items',
+});