+++ angular-view/src/app/service/server/word-items.service.ts
@@ -3,7 +3,7 @@ import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { catchError, tap } from 'rxjs/operators';
 
-import { UrlService } from './url.service';
+import { URL } from 'src/app/app.global-variable';
 import { HandleService } from './handle.service';
 
 import { Word } from '../../../../../common-code/models/word';
@@ -15,13 +15,12 @@ export class WordItemsService {
 
     constructor(
         private http: HttpClient,
-        private urlService: UrlService,
         private handleService: HandleService) {
     }
 
     get(): Observable<any> {
 
-        return this.http.get<any>(this.urlService.api_wordItems)
+        return this.http.get<any>(URL.API_WORDITEMS)
             .pipe(
                 tap(_ => this.handleService.log('fetched WordItems')),
                 catchError(this.handleService.handleError<any>('WordItems get'))
@@ -30,7 +29,7 @@ export class WordItemsService {
 
     post(data: Array<Word>): Observable<any> {
 
-        return this.http.post<any>(this.urlService.api_wordItems, data, this.handleService.httpOptions).pipe(
+        return this.http.post<any>(URL.API_WORDITEMS, data, this.handleService.httpOptions).pipe(
             tap((newData: any) => this.handleService.log(`added WordItems w/ id=${newData?.id}`)),
             catchError(this.handleService.handleError<any>('addHero'))
         );